backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
sentence-transformers>=2.0.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
PyJWT>=2.0.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
